Matchmaking with basic shuffle
